Add links to view model architecture with netron
This will add links to the documentation which allows us to visualize the networks with netron in the browser. For this to work I needed to use a CORS proxy to workaround the default CORS policy set by GitHub pages. See https://github.com/lutzroeder/netron/issues/312

Note that for BinaryNet and XNOR-Net it might take quite long to open the website since it needs to first download the model file which also include all the weights.